Implement live interpreter for OneFrame
With this change OneFrameDummy is replaced with the newly created
OneFrameLive. When getting a request, OneFrameLive will check in its
cache for rate to return. If the info in the cache is too old, or not
found, a request will be made to the external library to fetch all
rates data. The expiration time of the cache can be set in the
application.conf.

Moreover, some error handling has been made such as when
1. a unsupported currency name is given in either query parameter
2. if the OneFrame result fails to decode properly in OneFrameLive

Finally, Currency has been refactored to utilize enumeratum, which
eases the creation of more supported currencies if needed.